requirements.txt
Flask
Flask-Cors
gspread
requests
oauth2client
python-docx